Restructure vue components folders
