+++ frontend/src/components/common/UserScore.vue
